trusttech

+++ backend/app/Http/Controllers/ClientsController.php
@@ -28,39 +28,58 @@ class ClientsController extends Controller
         return response()->json(['message' => 'Client added successfully'], 201);
     }
 
-    public function updateClient(Request $request, $id)
+    public function updateClient(Request $request)
     {
         $validatedData = $request->validate([
             'nom' => 'required|string|max:255',
-            'email' => 'required|email|unique:clients,email,' . $id,
+            'email' => 'required|email',
             'telephone' => 'required|string|max:20',
             'societe' => 'required|string|max:255',
             'adresse' => 'required|string|max:255',
         ]);
 
-        $client = Clients::findOrFail($id);
-        $client->nom = $validatedData['nom'];
-        $client->email = $validatedData['email'];
-        $client->telephone = $validatedData['telephone'];
-        $client->societe = $validatedData['societe'];
-        $client->adresse = $validatedData['adresse'];
-        $client->save();
+        $client = Clients::where('nom', $validatedData['nom'])->first();
+        if (!$client) {
+            return response()->json(['error' => 'Client not found'], 404);
+        }
 
-        return response()->json(['message' => 'Client updated successfully']);
+        $client->update($validatedData);
+
+        return response()->json(['message' => 'Client updated successfully']);;
     }
 
-    public function deleteClient($id)
+    public function deleteClient(Request $request)
     {
-        $client = Clients::findOrFail($id);
+        $validatedData = $request->validate([
+            'nom' => 'required|string|max:255',
+        ]);
+
+        $client = Clients::where('nom', $validatedData['nom'])->first();
+        if (!$client) {
+            return response()->json(['error' => 'Client not found'], 404);
+        }
+
         $client->delete();
 
         return response()->json(['message' => 'Client deleted successfully']);
     }
 
-    public function getClientDetails($id)
+    public function getClientDetails(Request $request)
     {
-        $client = Clients::findOrFail($id);
+        $validatedData = $request->validate([
+            'nom' => 'required|string|max:255',
+        ]);
 
-        return response()->json(['client' => $client]);
+        $client = Clients::where('nom', $validatedData['nom'])->first();
+        if (!$client) {
+            return response()->json(['error' => 'Client not found'], 404);
+        }
+
+        return response()->json(['Client' => $client]);
+    }
+    public function getAllClients()
+    {
+        $client = Clients::all();
+        return response()->json(['Clients' => $client]);
     }
 }
